# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Employment Service API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  job_seeker: 'job_seeker',
  employer: 'employer',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserRegistrationRole = typeof UserRegistrationRole[keyof typeof UserRegistrationRole];


export const UserRegistrationRole = {
  job_seeker: 'job_seeker',
  employer: 'employer',
} as const;

export interface UserRegistration {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
  role: UserRegistrationRole;
}

export interface UserLogin {
  email: string;
  password: string;
}

export type JobType = typeof JobType[keyof typeof JobType];


export const JobType = {
  full_time: 'full_time',
  part_time: 'part_time',
  contract: 'contract',
  internship: 'internship',
  remote: 'remote',
} as const;

export interface Job {
  id: number;
  title: string;
  companyName: string;
  /** @nullable */
  companyLogo?: string | null;
  location: string;
  type: JobType;
  /** @nullable */
  salary: string | null;
  description: string;
  category: string;
  /** @nullable */
  requirements?: string | null;
  isFeatured?: boolean;
  createdAt: string;
}

export type JobInputType = typeof JobInputType[keyof typeof JobInputType];


export const JobInputType = {
  full_time: 'full_time',
  part_time: 'part_time',
  contract: 'contract',
  internship: 'internship',
  remote: 'remote',
} as const;

export interface JobInput {
  title: string;
  companyName: string;
  companyLogo?: string;
  location: string;
  type: JobInputType;
  salary?: string;
  description: string;
  category: string;
  requirements?: string;
  isFeatured?: boolean;
}

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  location?: string | null;
  openPositions: number;
}

export type InternshipApplicationStatus = typeof InternshipApplicationStatus[keyof typeof InternshipApplicationStatus];


export const InternshipApplicationStatus = {
  pending: 'pending',
  reviewing: 'reviewing',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface InternshipApplication {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  university: string;
  fieldOfStudy: string;
  /** @nullable */
  yearOfStudy?: string | null;
  /** @nullable */
  cvUrl?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  startDate: string;
  status: InternshipApplicationStatus;
  createdAt: string;
}

export interface InternshipApplicationInput {
  fullName: string;
  email: string;
  phone?: string;
  university: string;
  fieldOfStudy: string;
  yearOfStudy?: string;
  cvUrl?: string;
  coverLetter?: string;
  startDate: string;
}

export type ShaqotagApplicationStatus = typeof ShaqotagApplicationStatus[keyof typeof ShaqotagApplicationStatus];


export const ShaqotagApplicationStatus = {
  pending: 'pending',
  reviewing: 'reviewing',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface ShaqotagApplication {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  occupation: string;
  experience: string;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  cvUrl?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  preferredSector?: string | null;
  status: ShaqotagApplicationStatus;
  createdAt: string;
}

export interface ShaqotagApplicationInput {
  fullName: string;
  email: string;
  phone?: string;
  occupation: string;
  experience: string;
  skills?: string;
  cvUrl?: string;
  coverLetter?: string;
  preferredSector?: string;
}

export interface MembershipPlan {
  id: number;
  name: string;
  price: number;
  duration: string;
  features: string[];
  isPopular?: boolean;
}

export interface Membership {
  id: number;
  planId: number;
  fullName: string;
  email: string;
  phone: string;
  status: string;
  createdAt: string;
}

export interface MembershipInput {
  planId: number;
  fullName: string;
  email: string;
  phone: string;
}

export interface SiteSummary {
  totalJobs: number;
  totalCompanies: number;
  totalMembers: number;
  totalApplications: number;
}

export type ListJobsParams = {
search?: string;
category?: string;
type?: string;
location?: string;
limit?: number;
offset?: number;
};