# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Employment Service API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  location?: string | null;
  openPositions: number;
}