# lib/api-zod/src/generated/types/internshipApplicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Employment Service API
 * OpenAPI spec version: 0.1.0
 */

export interface InternshipApplicationInput {
  fullName: string;
  email: string;
  phone?: string;
  university: string;
  fieldOfStudy: string;
  yearOfStudy?: string;
  cvUrl?: string;
  coverLetter?: string;
  startDate: string;
}