# lib/api-zod/src/generated/types/membership.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Employment Service API
 * OpenAPI spec version: 0.1.0
 */

export interface Membership {
  id: number;
  planId: number;
  fullName: string;
  email: string;
  phone: string;
  status: string;
  createdAt: string;
}