# lib/api-zod/src/generated/types/membershipInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Employment Service API
 * OpenAPI spec version: 0.1.0
 */

export interface MembershipInput {
  planId: number;
  fullName: string;
  email: string;
  phone: string;
}