# lib/api-zod/src/generated/types/membershipPlan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Employment Service API
 * OpenAPI spec version: 0.1.0
 */

export interface MembershipPlan {
  id: number;
  name: string;
  price: number;
  duration: string;
  features: string[];
  isPopular?: boolean;
}