# lib/api-zod/src/generated/types/shaqotagApplicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Employment Service API
 * OpenAPI spec version: 0.1.0
 */

export interface ShaqotagApplicationInput {
  fullName: string;
  email: string;
  phone?: string;
  occupation: string;
  experience: string;
  skills?: string;
  cvUrl?: string;
  coverLetter?: string;
  preferredSector?: string;
}